pyproject.toml
[project]
name = "bricklaying-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame>=2.6.1",
    "pyyaml>=6.0.2",
    "types-pyyaml>=6.0.12.20250516",
]
